Move remaining ListViewUtils code to abstract base class AppListActivity.
